refactor(user): Extract the better-auth reads into AuthIdentityDao
Changes:
- Add AuthIdentityDao holding findByUserIds, deleteById and the timestamp coercion
- Inject it into UserService and drop the direct EntityManager dependency
- Remove the now-unused Timestamp, HashMap, OffsetDateTime and persistence imports

Reading and deleting rows in better-auth's `user` / `session` tables is a
different concern from the app_user profile service: that schema is owned by
the frontend's auth library and has no JPA entities, so it needs native
queries. Keeping it inline left UserService at 316 lines with two raw SQL
statements sitting next to the profile logic.

Notes:
- Review finding 34 (reviews/REVIEW-2026-07-24-DEV-VS-MAIN.md)
- UserService is now 291 lines, still above the AGENTS.md 50-100 target. The
  remaining bulk is genuine profile logic and would need a separate split.
